+++ .php_cs.dist
@@ -6,6 +6,7 @@ $config
     ->setUsingCache(true)
     ->getFinder()
     ->in(__DIR__)
+    ->exclude('_dev')
     ->exclude('vendor');
 
 return $config;
