+++ src/PayPal/Sdk/PayPalSdkConfigurationBuilder.php
@@ -18,7 +18,7 @@
  * @license   https://opensource.org/licenses/AFL-3.0 Academic Free License version 3.0
  */
 
-namespace PrestaShop\Module\PrestashopCheckout\Builder\PaypalSdkConfiguration;
+namespace PrestaShop\Module\PrestashopCheckout\PayPal\Sdk;
 
 use PrestaShop\Module\PrestashopCheckout\Environment\PaypalEnv;
 use PrestaShop\Module\PrestashopCheckout\ExpressCheckout\ExpressCheckoutConfiguration;
